+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v0.3
+- Now has a progress form and update log display using the wonderful work done by docdoom and T3KNOGHO57 from the ComicRack for the progress form and the log display.
+
+v0.2.5
+- Fix for unpopulated pages on setups without embedded xmls
+
 v0.2.2
 - Fix to make compatible with "Create a new Combined Script"